+++ python/ray/data/_internal/logical/rules/__init__.py
@@ -1,4 +1,4 @@
-from ray.data._internal.logical.rules.operator_fusion import OperatorFusionRule
+from ray.data._internal.logical.rules.operator_fusion import FuseOperators
 from ray.data._internal.logical.rules.randomize_blocks import ReorderRandomizeBlocksRule
 
-__all__ = ["ReorderRandomizeBlocksRule", "OperatorFusionRule"]
+__all__ = ["ReorderRandomizeBlocksRule", "FuseOperators"]
